server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
joblib==1.4.2
scikit-learn==1.6.1
python-dotenv==1.0.1
pymongo==4.12.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
cloudinary==1.44.1
